Split out client side infra to reduce transitive server side specific dependencies pulled into blazor webasm
