+++ src/AdminAssistant.Infra.ClientSide/Providers/LoggingProvider.cs
@@ -2,24 +2,7 @@ using System.Runtime.CompilerServices;
 using Microsoft.Extensions.Logging;
 
 namespace AdminAssistant.Infra.Providers;
-
-internal class ClientSideLoggingProvider : LoggingProvider, ILoggingProvider
-{
-    public ClientSideLoggingProvider(ILoggerFactory loggerFactory)
-        : base(loggerFactory, ILoggingProvider.ClientSideLogCategory)
-    {
-    }
-}
-
-internal class ServerSideLoggingProvider : LoggingProvider, ILoggingProvider
-{
-    public ServerSideLoggingProvider(ILoggerFactory loggerFactory)
-        : base(loggerFactory, ILoggingProvider.ServerSideLogCategory)
-    {
-    }
-}
-
-internal abstract class LoggingProvider : ILoggingProvider
+public abstract class LoggingProvider : ILoggingProvider
 {
     private readonly ILogger _logger;
