+++ src/AdminAssistant.Infra/Providers/ServerSideLoggingProvider.cs
@@ -1,0 +1,12 @@
+using Microsoft.Extensions.Logging;
+
+namespace AdminAssistant.Infra.Providers
+{
+    internal class ServerSideLoggingProvider : LoggingProvider, ILoggingProvider
+    {
+        public ServerSideLoggingProvider(ILoggerFactory loggerFactory)
+            : base(loggerFactory, ILoggingProvider.ServerSideLogCategory)
+        {
+        }
+    }
+}
