+++ src/AdminAssistant.sln
@@ -31,6 +31,8 @@ Project("{E53339B2-1760-4266-BCC7-CA923CBCF16C}") = "docker-compose", "docker-co
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AdminAssistant.Retro", "AdminAssistant.Retro\AdminAssistant.Retro.csproj", "{E72BBBA7-8A75-47E2-BFF4-7460163616E2}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AdminAssistant.Infra.ClientSide", "AdminAssistant.Infra.ClientSide\AdminAssistant.Infra.ClientSide.csproj", "{CA7C589A-4830-495B-9B96-B2E9D01E78F4}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -77,6 +79,10 @@ Global
 		{E72BBBA7-8A75-47E2-BFF4-7460163616E2}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{E72BBBA7-8A75-47E2-BFF4-7460163616E2}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{E72BBBA7-8A75-47E2-BFF4-7460163616E2}.Release|Any CPU.Build.0 = Release|Any CPU
+		{CA7C589A-4830-495B-9B96-B2E9D01E78F4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{CA7C589A-4830-495B-9B96-B2E9D01E78F4}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{CA7C589A-4830-495B-9B96-B2E9D01E78F4}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{CA7C589A-4830-495B-9B96-B2E9D01E78F4}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
